sha1-file: convert some static variables to local
sha1-file.c has many static variables inside functions which make then
non-reentrant and non-thread-safe. As an effort to allow threading on
the pack access code, let's convert some of this variables to local.

The functions 'cleaned' in this first batch are:
- fill_loose_path
- check_and_freshen_odb
- stat_loose_object
- oid_object_info_extended
- write_loose_object
- open_loose_object

And the following had to be adjusted, by consequence:
- map_loose_object_1
- loose_object_info

Note: this is a refactor-only patch with no additions.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>