object-store: allow threaded access to object reading
Allow object reading to be performed by multiple threads protecting it
with a set of locks. The locks usage can be toggled with
enable_obj_read_locks() and disable_obj_read_locks(). Currently, the
functions which can be safely called in parallel are:
read_object_file_extended(), repo_read_object_file(),
read_object_file(), read_object_with_reference(), read_object(),
oid_object_info() and oid_object_info_extended().

Probably there are many spots in these functions that could be executed
unlocked (and thus, in parallel). But, for now, we are most interested
in allowing parallel access to zlib inflation. This is one of the
sections where object reading spends most of the time and it's already
thread-safe. So, to take advantage of that, the respective lock is
released when entering it and re-acquired right after. We may refine the
locks to also exploit other possible parallel spots in the future, but
threaded zlib inflation should already give great speedups.

Note that add_delta_base_cache() was also modified to skip adding
already present entries to the cache. This wasn't possible before, but
now it is since phase I and phase III of unpack_entry() may execute
concurrently.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>