grep: remove locks on object reading operations
git-grep has locks around the object reading operations it performs. But
these have an internal lock now, which can be enabled through
enable_obj_read_lock(). So let's use it and drop git-grep's ones. To do
so, we also need to protect the packed_git_mru list of struct
raw_object_store and the windows of struct packed_git. Finally, for the
textconv operations on git-grep, a simple mutex arround it is enough.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>