grep: move driver pre-load out of critical section
In add_work() we pre-load the userdiff drivers before adding the
grep_source in the todo list. This operation is currently being
performed after acquiring the grep_mutex but, as it's thread-safe, we
don't need to protect it here. So let's move it out of the critical
section which should avoid thread contention and improve performance.

Running[1] `git grep --threads=8 abcd[02] HEAD` on chromium's
repository[2], I got the following mean times for 30 executions after 2
warmups:

        Original         |  6.2093s
-------------------------|-----------
 Out of critical section |  5.7532s

[1]: Tests performed on an i7-7700HQ with 16GB of RAM and SSD.
[2]: chromium’s repo at commit 03ae96f (“Add filters testing at DSF=2”,
         04-06-2019), after a 'git gc' execution.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>