config: add repo_config_with_options()
Because config_with_options() relies on the global the_repository, a
callee that wants to work on a subrepo must first add it to alternates.
This can negatively affect performance and memory. So let's rename and
modify it to take a struct repository as parameter and pass it down in
the call stack. Also add an alias config_with_options() which uses the
global the_repository by default, for backwards compatibility.

The following patch will take care of actually using the added function
in place of config_with_options() at the callee which is handling a
subrepo.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>