submodule-config: don't add subrepo to alternates
It was previously necessary to add a subrepo to the alternates list of
the_repository because config_with_options() would work in it by
default. But there's now a repo_config_with_options() which accepts a
struct repository as a parameter. So use it and remove the addition to
the alternates. This might increase performance and/or free memory. A
future patch will also use this change to try improving parallelization
in git-grep, which indirectly calls this function. As it doesn't rely on
the_repository, anymore, it should be possible to remove some locks at
git-grep without the chance of race conditions.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>