object: allow parse_object_or_die() to handle any repo
parse_object_or_die() was added before parse_object() learned to handle
arbitrary repositories. Now that it did, let's also bring this behavior
improvement to the former, adding a struct repository argument. Also
adjust the callers and, while we are here, avoid using `oid_to_hex(oid)`
for the third parameter as it's the default when NULL is given.

This will be used in the following patch, to allow git-grep to parse
submodule's objects.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>