grep: don't add submodules to the alternates list
When grepping with --recurse-submodules, the object directory of the
submodule is added to the alternates list. This makes Git need to watch
out for more packfiles and also complicates the work of optimizing
threads (we need to implement mutual exclusion between the list adding
and object reading). Now that object store is a member of the struct
repository, it's possible to use the submodules' instance directly,
without the need to add the objects to the alternates list. So let's
pass the subrepo down to the threads and replace function calls which
use the_repository by default for their more general counterparts.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>