grep: move parse_object_or_die() into critical section
The call to parse_object_or_die(), at grep_submodule() is outside the
critical section of grep_read_mutex, which could lead to race conditions
when grepping the index. That's not a big deal, today, because threading
is disabled for this case due to low performance. But planning to
re-enable them in the future, it's good to already take care of this so
that we don't get to this problem latter.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>