completion: provide a function for csv completion
Some Git cli options accept a comma separated list of values. I.e.
something in the format of `--option=val1[,val2[,...]]`. However, there
is currently no easy way to provide completion for such options. Let's
add a helper function for this task. The function will be used in the
following patch to add completion for a git-diff option that takes csv
values.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>