config: load the correct config.worktree file
One of the steps in do_git_config_sequence() is to load the
worktree-specific config file. Although the function receives a git_dir
string, it relies on git_pathdup(), which uses the_repository->git_dir,
to make the path to the file. Thus, when a submodule has a worktree
setting, a command executed in the superproject that recurses into the
submodule won't find the said setting. Such scenario might not be needed
now, but it will be in the following patch. git-grep will learn to honor
sparse-checkouts and, running with --recurse-submodules, the submodule's
sparse-checkout settings must be loaded. As these settings are stored in
the config.worktree file, they would be ignored without this patch.

The fix is simple, we replace git_pathdup() with mkpathdup(), to format
the path with the given git_dir. This is the same idea used to make the
config.worktree path in setup.c:check_repository_format_gently().

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>