entry: extract thread-unsafe ops from write_entry()
In order to implement parallel checkout, in the upcoming patches, we
should be able to call write_entry() within threads. However, there are
two non-thread-safe operations performed by it: the attributes lookup
and the cache entry update after writing it to disk. In preparation for
the threaded calls, extract those two operations out of write_entry().

Note: this allows us to call the function in parallel [1], as long as
the entry being written does not require an external process filter
(which we will make sure to check in the upcoming patch where this will
be implemented).

[WIP] Provisory note to myself: [1] is true assuming
[die|error|warning]_errno() and oid_to_hex() are thread-safe, which they
are currently not. So, on errors, we might still get race conditions,
although the actual risk must be pretty small. A *lot* of places may
call this functions on error, so it might be difficult to just replace
them all with the reentrant variants. But I'm working on making those
functions thread-safe in a separate series. We probably want to make
sure the problem is solved  before integrating this one.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>