unpack-trees: add basic support for parallel checkout
This new interface allows us to enqueue some of the entries being
checked-out to later call write_entry() for them in parallel. For now,
the parallel checkout machinery is enabled by default and there is no
user configuration, but run_parallel_checkout() just writes the queued
entries in sequence. The next patch will actually implement the
parallelism and make it configurable.

Note that, before calling write_entry(), checkout_entry() checks that
the path is not already present in the working tree (and if so, the code
also queries whether it's up-to-date). These checks are important for
two reasons: (1) to avoid re-writing files that are not changed when,
for example, switching branches; and (2) to detect and handle path
collisions among the files being checked-out (which might happen, for
example, with case-sensitive files in case-insensitive file systems).
Now that we are enqueuing the entries to call write_entry() later, (1)
will still work correctly but not (2). This is because we won't check
for the paths' existence before writing each queued entry.
write_entry() assumes checkout_entry() has already done that and, in
fact, by the time checkout_entry() was called, the path might not be
present (as the entry was queued for writing), but it might exist now.
This reasoning is even clearer when we add threads, as they will compete
to write paths that collide. We handle this case by detecting which
entries made write_entry() fail due to path collision, and later call
checkout_entry() again for the said entries, sequentially. This way we
have no overhead when there aren't collisions. Also, feeding the
colliding entries sequentially (and in order) to checkout_entry()
reproduces the currently serial behavior. So we should get deterministic
results even though threads might compete for some paths. Also, this
mechanism is the same used by delayed checkout, so there should be no
extra risks here.

Note: we continue the loop of write_entry() anyway even if the previous
call returns an error. This is how checkout_entry() is called in
builtin/checkout.c:checkout_paths() and unpack-trees.c:check-updates().
In the case of fatal errors, die() aborts the loop.

Co-authored-by: Duy Nguyen <pclouds@gmail.com>
Co-authored-by: Jeff Hostetler <jeffhost@microsoft.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>