unpack-trees: add basic support for parallel checkout
This new interface allows us to enqueue some of the entries being
checked out to later uncompress, smudge, and write them in parallel. For
now, the parallel checkout machinery is enabled by default and there is
no user configuration, but run_parallel_checkout() just writes the
queued entries in sequence (without spawning additional workers). The
next patch will actually implement the parallelism and, later, we will
make it configurable.

Note that not all entries are eligible for parallel checkout. More
specifically, we exclude regular files that require external filters and
symlinks. The former is excluded because we can't guarantee that such
filters are parallel-process safe, and the latter because, it could
cause race conditions in case of collisions with the leading paths of
other entries. The ineligible entries are checked out sequentially
*before* the eligible ones, using the classic code.

Speaking of path collisions (which can happen e.g. when checking out
case-sensitive files in case-insensitive file systems), parallel
checkout uses two complementary methods to detect when entries collide
on disk:

- If the collision was at basename (e.g. 'a/b' and 'a/B'), the framework
  detects it by looking for EEXIST and EISDIR errors after an
  open(O_CREAT | O_EXCL) failure.

- If the collision was at dirname (e.g. 'a/b' and 'A'), it is detected
  at the has_dirs_only_path() check, which is done for the leading path
  of each item in the parallel checkout queue.

Once we add parallelism, these verifications will ensure that we don't
racily write to different paths that represent the same file on disk.
Both verifications rely on the fact that, before enqueueing an entry for
parallel checkout, checkout_entry() makes sure that there is no file at
the entry's path and that its leading components are all real existing
directories. Therefore if the framework later finds out that these
conditions changes, it indicates that there was a collision (either
between two parallel-eligible entries or between an eligible and an
ineligible entry).

When a collision is detected, the parallel_checkout_item is marked with
PC_ITEM_COLLIDED and skipped. Later, these items are sequentially fed to
checkout_entry() again. This is similar to the way the current code
deals with collisions, overwriting the previously checked out entries
with the subsequent ones. The only difference is that, since we no
longer create the files in the same order that they appear on index, we
are not able to determine which of the colliding entries will survive on
disk (for the current code, it is always the last one).

Alternatively, we could possibly skip the additional checkout_entry()
calls and don't overwrite already written entries. This idea seemed to
work well in simple tests, but because just one entry of each colliding
group would be actually written, the others would have null lstat()
fields on the index. As a consequence, all subsequent Git commands that
need to read and refresh the index would have to go to the file system
to check that the unwritten entries are dirty. For a group of N
colliding entries, this means reading, converting and hashing the
written file N - 1 times. By checking out all colliding entries, like
the current code does, we only pay the overhead once, at checkout.

Co-authored-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Co-authored-by: Jeff Hostetler <jeffhost@microsoft.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>